The PipelineExecutionEstimator allows the caller to chain multiple Estimators end-to-end to form a pipeline (or DAG). During training, data trickles down to the currently untrained Estimator within the chain. Once all training is complete, a Transformer is created that only invokes the Transformers associated with TransformerEstimators in the original chain.

This code attempts to provide compile-time warnings when Estimators are chained together in an incompatible way.

+++ src/FeaturizerPrep/UnitTests/Traits_UnitTest.cpp
@@ -141,8 +141,8 @@ TEST_CASE("Transformer_Vectors") {
 
 TEST_CASE("Transformer_Maps") {
     std::map<std::int16_t, std::double_t> m;
-    m.insert(std::pair<std::int16_t, std::double_t>(5, 35.8));
-    m.insert(std::pair<std::int16_t, std::double_t>(93, 0.147));
+    m.insert(std::pair<std::int16_t, std::double_t>(static_cast<std::int16_t>(5), 35.8));
+    m.insert(std::pair<std::int16_t, std::double_t>(static_cast<std::int16_t>(93), 0.147));
     std::string map_res = Traits<std::map<std::int16_t, std::double_t>>::ToString(m);
     std::string map_s{ "{5:35.8,93:0.147}" };
 	CHECK(map_res == map_s);
